Move menu bar to main process
